Add files via upload
A grammar-based blackbox fuzzer to test SQLite, which is the most widely deployed SQL database engine in the world. IMPORTANT: The page at https://www.sqlite.org/lang.html lists all commands understood by SQLite. This page guides the grammar implementation. SQLite is implemented in C and for this project, Linux was used.
this project makes use of the ProbabalisticGrammarFuzzer from thefuzzingbook library in python to generate fuzzed inputs for the SQL engine.